+++ bigml/tests/create_script_steps.py
@@ -19,7 +19,7 @@ import time
 import json
 import os
 from datetime import datetime
-from world import world, logged_wait
+from world import world, logged_wait, res_filename
 from nose.tools import eq_, assert_less
 
 from bigml.api import HTTP_CREATED
@@ -27,6 +27,7 @@ from bigml.api import HTTP_ACCEPTED
 from bigml.api import FINISHED
 from bigml.api import FAULTY
 from bigml.api import get_status
+from bigml.util import is_url
 
 from read_script_steps import i_get_the_script
 
@@ -41,6 +42,18 @@ def the_script_code_and_attributes(step, source_code, param, param_value):
 
 #@step(r'I create a whizzml script from a excerpt of code "(.*)"$')
 def i_create_a_script(step, source_code):
+    resource = world.api.create_script(source_code)
+    world.status = resource['code']
+    eq_(world.status, HTTP_CREATED)
+    world.location = resource['location']
+    world.script = resource['object']
+    world.scripts.append(resource['resource'])
+
+
+#@step(r'I create a whizzml script from file "(.*)"$')
+def i_create_a_script_from_file_or_url(step, source_code):
+    if not is_url(source_code):
+        source_code = res_filename(source_code)
     resource = world.api.create_script(source_code)
     world.status = resource['code']
     eq_(world.status, HTTP_CREATED)